Add Fluvio-test self check

+++ crates/fluvio-test/src/main.rs
@@ -1,4 +1,5 @@
 use std::process::exit;
+use std::env;
 use structopt::StructOpt;
 use fluvio::Fluvio;
 use fluvio_test_util::test_meta::{BaseCli, TestCase, TestCli, TestOption};
@@ -95,7 +96,12 @@ fn run_test(
         Err(_) => {
             // nix uses pid 0 to refer to the group process, so reap the child processes
             let pid = Pid::from_raw(0);
-            kill(pid, Signal::SIGTERM).expect("Unable to kill test process");
+
+            // Only send SIGTERM outside of CI (where CI env var is unset), because it might cause job to report as cancelled
+            if env::var("CI").is_err() {
+                kill(pid, Signal::SIGTERM).expect("Unable to kill test process");
+            }
+
             exit(1);
         }
     }
